MJ Adding edited files of guess who game
The “guess who” button wasn’t working so I added a script that it was
missing. Plus, some of the files had some “promise” errors in the
console due to the use of mouseover and mouseout so I changed those as
well so there wouldn’t be errors in the console.